Moved logic into auth service
