+++ MyMoney.Web/ClientApp/src/app/shared/constants.ts
@@ -1,1 +1,3 @@
 export const SESSION_LOCAL_STORAGE_KEY = 'my-money:session';
+export const LOGIN_PAGE_PATH = 'login';
+export const REGISTER_PAGE_PATH = 'register';
