+++ MyMoney.Web/ClientApp/src/app/shared/services/authentication-service.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { first, map } from 'rxjs/operators';
 import { ILoginResultDto, IRegisterDto } from '../api';
 import { Store } from '@ngrx/store';
 import { IAppState } from '../state/app-state';
@@ -9,6 +9,7 @@ import { selectCurrentSession } from '../state/selectors/session.selector';
 import { ISessionModel } from '../state/types';
 import { Router } from '@angular/router';
 import { UserApi } from '../api/user.api';
+import { SESSION_LOCAL_STORAGE_KEY } from '../constants';
 
 @Injectable({ providedIn: 'root' })
 export class AuthenticationService {
@@ -45,5 +46,46 @@ export class AuthenticationService {
    public logout(): void {
       this.store.dispatch(new ClearSessionAction());
       this.router.navigate(['/login']);
+   }
+
+   public checkSession(): Observable<boolean> {
+      return this.store.select(selectCurrentSession).pipe(map((session: ISessionModel | null) => {
+
+         if (this.IsValid(session)) {
+            return true;
+         }
+
+         try {
+            const sessionData: string | null = localStorage.getItem(SESSION_LOCAL_STORAGE_KEY);
+
+            if (sessionData !== null) {
+
+               const browserSession: ISessionModel = JSON.parse(sessionData);
+
+               if (this.IsValid(browserSession)) {
+                  console.log('Session: Using cached session from local storage');
+                  this.store.dispatch(new StartSessionAction(browserSession.token, browserSession.sessionEnd));
+                  return true;
+               }
+            }
+         } catch (error) {
+            console.log(error);
+         }
+
+         this.store.dispatch(new ClearSessionAction());
+         return false;
+      }));
+   }
+
+   private IsValid(session: ISessionModel | null): boolean {
+
+      if (session === null) {
+         return false;
+      }
+
+      const now = new Date(Date.now()).getTime();
+      const validTo = Date.parse(session.sessionEnd);
+
+      return validTo > now;
    }
 }
